requirements.txt
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.1.2
PyYAML==6.0
setuptools==62.3.4
wheel==0.37.1
